# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AINA Knowledge Scraper v2 API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /drafts:
    get:
      operationId: listDrafts
      tags: [drafts]
      summary: List drafts (contributor sees own, admin sees all)
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, submitted, approved, rejected]
        - name: source_type
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of drafts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DraftList"

    post:
      operationId: createDraft
      tags: [drafts]
      summary: Create a new draft
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DraftInput"
      responses:
        "201":
          description: Draft created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Draft"
  /drafts/{id}:
    get:
      operationId: getDraft
      tags: [drafts]
      summary: Get a single draft
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Draft detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Draft"
        "404":
          description: Not found

    patch:
      operationId: updateDraft
      tags: [drafts]
      summary: Update a draft
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DraftUpdate"
      responses:
        "200":
          description: Updated draft
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Draft"

    delete:
      operationId: deleteDraft
      tags: [drafts]
      summary: Delete a draft
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /drafts/{id}/submit:
    post:
      operationId: submitDraft
      tags: [drafts]
      summary: Submit draft for admin review
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Draft submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Draft"
  /drafts/{id}/approve:
    post:
      operationId: approveDraft
      tags: [drafts]
      summary: Approve draft and push to knowledge base (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Draft approved and moved to knowledge base
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Draft"
  /drafts/{id}/reject:
    post:
      operationId: rejectDraft
      tags: [drafts]
      summary: Reject a draft with a reason (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectInput"
      responses:
        "200":
          description: Draft rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Draft"

  # ── Scrape / AI Processing ───────────────────────────────────────────────
...
components:
  schemas:
...
    Draft:
      type: object
      required: [id, title, content, status, submitted_by, created_at, source_type, relevance_score]
      properties:
        id:
          type: string
        title:
          type: string
        content:
          type: string
        summary:
          type: ["string", "null"]
        tags:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        source_url:
          type: ["string", "null"]
        source_type:
          type: string
          enum: [url, manual, pdf, instagram, auto]
        relevance_score:
          type: integer
          minimum: 0
          maximum: 100
        status:
          type: string
          enum: [draft, submitted, approved, rejected]
        submitted_by:
          type: string
        rejection_reason:
          type: ["string", "null"]
        created_at:
          type: string
          format: date-time
    DraftList:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Draft"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    DraftInput:
      type: object
      required: [title, content, source_type]
      properties:
        title:
          type: string
        content:
          type: string
        summary:
          type: string
        tags:
          type: string
        category:
          type: string
        source_url:
          type: string
        source_type:
          type: string
          enum: [url, manual, pdf, instagram, auto]
        relevance_score:
          type: integer
    DraftUpdate:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        summary:
          type: string
        tags:
          type: string
        category:
          type: string
    RejectInput:
      type: object
      required: [rejection_reason]
      properties:
        rejection_reason:
          type: string

    # ── Scrape inputs ──────────────────────────────────────────────────────
...